% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ampCounter.R
\name{enumerateAmplifications}
\alias{enumerateAmplifications}
\title{Enumerate the multiple strand displacement fragments generated given a set of primers}
\usage{
enumerateAmplifications(forwards, reverses, strand = "+", maxLength = 50000,
  minLength = 1, maxTotalLength = Inf, fragmentStart = 1,
  fragmentEnd = Inf, baseName = "", vocal = FALSE, previousLength = 0)
}
\arguments{
\item{forwards}{Sorted positions of the start of the primers landing sites on the forward strand of the target sequence (5'-most base in the genome plus strand)}

\item{reverses}{Sorted positions of the end of the primers landing sites on the reverse strand of the target sequence (3'-most base in the genome plus strand)}

\item{strand}{"+" or "-" for indicating what strand the target sequence originates from. Note if starting with double stranded fragments then the function should be run once with "+" and once with "-" and the results concatenated.}

\item{maxLength}{The expected max length fragment the polymerase can generate in one run.}

\item{minLength}{Discard fragments shorter than minLength (can help with run time if many short fragments are generated)}

\item{maxTotalLength}{Discard fragments for which the sum of all the ancestors bases are longer than maxTotalLength. If the polymerase has some probability of falling off each base amplified then it becomes less and less likely to reach a deeply amplified product. Using maxTotalLength to end the recursion once a certain total length is reach can help with run time.}

\item{fragmentStart}{The 5' end of the current fragment (for use in recursion more than manually)}

\item{fragmentEnd}{The 3' end of the current fragment (for use in recursion more than manually)}

\item{baseName}{Prefix for tracking fragment ancestors (for use in recursion more than manually)}

\item{vocal}{TRUE or FALSE Output progress?}

\item{previousLength}{How many bases need to be amplified to reach this fragment (for use in recursion more than manually)}
}
\value{
A data frame with columns start, end, strand, name, previousLength and length. Start and end and strand are the positions of 5' and 3' end on the appropriate strand of the target sequence. Name is a concatenation of the ancestors of that fragment. Previous length is the sum of the fragment lengths of the ancestors of this fragment. Length is the width of this fragment. Returns NULL if no amplifications.
}
\description{
Take two vectors of forward and reverse primer binding locations
generate a list of amplification products predicted to be
generated by multiple strand displacement amplification.
}
\examples{
enumerateAmplifications(c(10,20,30),c(40,50,60),maxLength=45)
}
\references{
\url{https://en.wikipedia.org/wiki/Multiple_displacement_amplification}
}
\seealso{
\code{\link{countAmplifications}}
}
\concept{
multiple strand displacement amplification fragment prediction recursive recursion #'
}

