% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ampCounter.R
\name{generateAmplificationTable}
\alias{generateAmplificationTable}
\title{Fast calculation of expected multiple strand displacement amplification}
\usage{
generateAmplificationTable(nForwards = 10, nReverses = 10)
}
\arguments{
\item{nForwards}{Calculate a matrix from zero to this many forward primers}

\item{nReverses}{Calculate a matrix from zero to this many reverse primers}
}
\value{
A nForwards+1 row, nReverses+1 column matrix of the expected amplifications for each combination of numbers
of forward and reverse primers. Row 1 is for 0 forward primers within range, row 2 is for 1 forward primer,
column 1 is for 0 reverse primers, column 2 is for 1 reverse primers ...
}
\description{
Uses dynamic programing to build a matrix of the expected amplifications for positions with a given a number
of forward primers upstream and reverse primers downstream and within range of the polymerase.
}
\examples{
generateAmplificationTable(20,20)
}
\seealso{
\code{\link{enumerateAmplifications}}
}
\concept{
multiple strand displacement amplification fragment prediction dynamic programming
}

