% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ampCounter.R
\name{generateRandomPrimers}
\alias{generateRandomPrimers}
\title{Generate random primers for testing}
\usage{
generateRandomPrimers(genomeSize, frequency)
}
\arguments{
\item{genomeSize}{Max position for primers}

\item{frequency}{About how much space between primers. Function will generate approximately genomeSize/frequency primers}
}
\value{
sorted list of positions
}
\description{
Generate random position for primers. Mostly for internal testing and examples
}
\examples{
ampcountr:::generateRandomPrimers(10000,100)
}
\keyword{internal}

