% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ampCounter.R
\name{plotFrags}
\alias{plotFrags}
\title{Plot fragments generated from multiple strand displacement}
\usage{
plotFrags(frags, label = TRUE)
}
\arguments{
\item{frags}{A data.frame with columns start, end, strand  (and name if label is TRUE) e.g. the output of \code{\link{enumerateAmplifications}}}

\item{label}{TRUE or FALSE if fragments should be labeled with name from the name column of frags}
}
\value{
NULL. Draws plot to current device
}
\description{
Take a data.frame with columns start, end, strand  (e.g. the output of \code{\link{enumerateAmplifications}}) and plot
an arrow representation of the fragments to the current device.
}
\examples{
frags<-enumerateAmplifications(c(10,20,30),c(40,50,60))
plotFrags(frags)
}
\seealso{
\code{\link{enumerateAmplifications}}
}
\concept{
multiple strand displacement amplification fragment plot arrows
}

