% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ampCounter.R
\name{predictAmplificationsSingleStrand}
\alias{predictAmplifications}
\alias{predictAmplificationsSingleStrand}
\title{Calculate expected multiple strand displacement for a series of forwards and reverse primers}
\usage{
predictAmplificationsSingleStrand(forwards, reverses, maxLength = 30000,
  genomeSize = max(c(forwards + maxLength - 1, reverses)))

predictAmplifications(forwards, reverses, maxLength = 30000,
  genomeSize = max(c(forwards + maxLength - 1, reverses)))
}
\arguments{
\item{forwards}{Leftmost positions of primer binding sites landing on the 5' strand (1-based)}

\item{reverses}{Rightmost positions of primer binding sites landing on the 3' strand (1-based)}

\item{maxLength}{Maximum length fragment generated by polymerase}

\item{genomeSize}{Maximum position possible for amplification}
}
\value{
A data frame with columns:
     \describe{
       \item{start:}{1-based start coordinate of region}
       \item{end:}{1-based end coordinate of region}
       \item{forwards:}{number of forward primers spanning the region (for predictAmplificationsSingleStrand)}
       \item{reverses:}{number of reverse primers spanning the region (for predictAmplificationsSingleStrand)}
       \item{amplifications:}{expected amplifcations for the region}
     }
}
\description{
Calculate the expected amplifications across a genome for a set of forward and reverse primer binding sites used in multiple strand displacement amplification
}
\section{Functions}{
\itemize{
\item \code{predictAmplifications}: Calculate the expected amplifications across a genome for a set of forward and reverse primer binding sites for both strands
}}
\examples{
predictAmplificationsSingleStrand(c(1,10,20),c(15,25,35),maxLength=40,genomeSize=45)
predictAmplifications(c(1,10,20),c(15,25,35),maxLength=40,genomeSize=45)
}
\concept{
multiple strand displacement amplification fragment prediction dynamic programming genome
}

