% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ampCounter.R
\docType{package}
\name{ampCounter}
\alias{ampCounter}
\alias{ampCounter-package}
\title{Functions to predict the amplification generated by a strand displacement amplification}
\description{
Given a list of forward and reverse primer binding sites predict the amplification that will result from a strand displacement reaction
}
\details{
The main functions are:
     \describe{
       \item{\code{\link{countAmplifications}}:}{to predict the number of amplifications for a single region}
       \item{\code{\link{predictAmplifications}}:}{to predict the number of amplifications across a whole genome}
       \item{\code{\link{enumerateAmplifications}}:}{to generate the predicted fragments for small sets of primers}
     }
}
\examples{
forwards<-ampCounter:::generateRandomPrimers(100000,1000)
reverses<-ampCounter:::generateRandomPrimers(100000,1000)
amplifications<-predictAmplifications(forwards,reverses,maxLength=10000)

plot(1,1,type='n',xlim=c(1,max(amplifications$end)),ylim=c(1,max(amplifications)),xlab='Genome position',ylab='Amplifications',log='y')
segments(amplifications$start,amplifications$amplification,amplifications$end,amplifications$amplification)
}
\seealso{
\url{https://en.wikipedia.org/wiki/Multiple_displacement_amplification}
}

