% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ampCounter.R
\docType{data}
\name{amplificationLookup}
\alias{amplificationLookup}
\title{Lookup table for predicted amplifcation}
\format{A 1001 row, 1001 column matrix of the expected amplifications for each combination of 0-1000
forward and reverse primers spanning a region. Row 1 is for 0 forward primers within range, row 2 is for 1 forward primer,
column 1 is for 0 reverse primers, column 2 is for 1 reverse primers ...}
\source{
generateAmplificationTable(1000,1000)
}
\usage{
amplificationLookup
}
\description{
A 1001x1001 matrix of predicted amplification for regions spanned by 0-1000 primers
}
\keyword{datasets}

